refactor: split analysisSemantics.ts (1122 lines) into 5 focused modules
Split pure logic file into:
- analysisSemantics.types.ts (95) — all types
- analysisSemantics.utils.ts (277) — constants + parsing helpers
- analysisSemantics.preference.ts (293) — preference model builders
- analysisSemantics.reliability.ts (300) — reliability model builders
- analysisSemantics.ts (230) — barrel + main entry functions

12/12 existing tests pass. Codex: 3/4 PASS (1 false alarm).
